requirements.txt
-i https://pypi.org/simple
alembic==1.11.1 ; python_version >= '3.7'
aniso8601==9.0.1
bcrypt==4.0.1 ; python_version >= '3.6'
blinker==1.6.2 ; python_version >= '3.7'
boto3==1.26.135
botocore==1.29.135 ; python_version >= '3.7'
click==8.1.3 ; python_version >= '3.7'
faker==18.9.0
flask==2.2.2
flask-bcrypt==1.0.1
flask-cors==3.0.10
flask-migrate==4.0.4
flask-restful==0.3.9
flask-sqlalchemy==3.0.3
greenlet==2.0.2 ; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==20.1.0
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
jmespath==1.0.1 ; python_version >= '3.7'
mako==1.2.4 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
psycopg2-binary==2.9.6
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.21.0
pytz==2023.3
s3transfer==0.6.1 ; python_version >= '3.7'
setuptools==67.7.2 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.13 ; python_version >= '3.7'
sqlalchemy-serializer==1.4.1
typing-extensions==4.5.0 ; python_version >= '3.7'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.3.4 ; python_version >= '3.8'
